Add set of chunks above the current position to predetermined positions.
Summary: Currently if you are standing on the top cell of a chunk, the chunk above you won't be rendered, which causes weird rendering to occur when you're walking just below sea level (because you're on the chunk below, but the land above isn't being rendered).  This adds the chunks above your position to the predetermined render list, which resolves the issue.

Test Plan: Run the game and walk into the sea.  The land should not disappear while you're walking around the seabed.

Reviewers: Sabre070

Reviewed By: Sabre070

CC: buildbot

Maniphest Tasks: T113

Differential Revision: http://code.redpointsoftware.com.au/D96