Replace flow bundles with predetermined ResultData struct.
Summary: This removes flow bundles in favour of a predetermined ResultData structure.  This change not only drastically eases memory requirements during generation, but reduces the average generation time in-game to around 40ms (in the Tychaia Tool it's reported as low as 17ms).

Test Plan: Apply the patch and observe the awesome speed!

Reviewers: Sabre070

Reviewed By: Sabre070

CC: buildbot

Maniphest Tasks: T112

Differential Revision: http://code.redpointsoftware.com.au/D97