Implement support for health, stamina and mana.
Beings (and subsequently players) all have health, stamina and mana
properties now.  The status bar displays the current player's health,
stamina and mana as percentage bars.  In addition, there's also a
"stats" command that allows you to set the health, stamina and mana
of the player since at the moment all statistics are set to zero
(including the maximums).

There's also some refactoring to bring it inline with StyleCop
requirements.