Upgrade Dx to version 2.0
Summary:
This brings in Dx 2.0 into the Tychaia repository, addressing tasks T119, T120 and T121.

It adds a workaround for DefaultTextureAtlasAssetFactory; it seems this can't generate the texture atlas when running with a network asset manager.  As to why I'm not too sure yet, but I spent a few hours and wasn't getting anywhere with it, so for now just catch the exception and silently fail (in-game the chunks just render with a black texture, but this only affects the game when running with the networked asset manager).

TychaiaProfiler has been updated to implement the Dx interface for profiling; Protogame is updated so that the implementation of IProfiler (TychaiaProfiler in our case) will be used for network asset manager profiling if it implements INetworkProfilerEndpoint.  This means that the graph in the top-left hand side will display the number of distributed events that occur in a frame as a cyan line.

Test Plan:
None of the changes should affect the game when running under normal circumstances; these changes only affect what happens when you run with the `--asset-manager` argument.

First and most importantly, make sure the game still runs and works perfectly fine after this patch is applied.  Second, it you run the game with `--asset-manager`, then the network asset manager should start, connect and the game should run (although when playing the game you'll notice the chunks render with a black texture due to the issue mentioned above).

If the game or network asset manager crashes when you run the game with `--asset-manager`, then that's an issue that needs addressing, but I'll probably address it when I profile the network asset manager for T70 anyway, so it shouldn't be considered a blocker for landing this revision (do make a note of it in the comments if you encounter crashes when running with `--asset-manager` though).

Reviewers: Sabre070, buildbot

CC: buildbot

Maniphest Tasks: T70, T119, T120, T121

Differential Revision: http://code.redpointsoftware.com.au/D101