Remove redundant ResultData and BlockInfo structures.
Summary:
This removes the redundant ResultData and BlockInfo structures in the procedural generation system, replacing them with the Cell data structure which is directly used in-game.

This does have a noticable effect on the amount of GC required; when moving around the map there's a lot less blocking GC events and the GC halts for less time.

Fix lint issues

Test Plan: Apply the patch and run the game.  On Windows you should barely notice any GC occurring at all, while on Linux you'll notice some GC but nowhere near as serious as it was before.

Reviewers: Sabre070, buildbot

CC: buildbot

Maniphest Tasks: T123

Differential Revision: http://code.redpointsoftware.com.au/D102