Move code from Tychaia to Tychaia.Runtime.
Summary:
This is required as part of the code migration described in T125.  We can't sanely implement a server-client model until we move all common code out into Tychaia.Runtime.

At the moment there's been a little bit of abstraction (with ClientRuntimeChunk) so that the game still works, but there's expected to be more code refactoring to support the server-client architecture in the near future.

Test Plan: Make sure the game still builds and runs correctly.

Reviewers: Sabre070

Reviewed By: Sabre070

CC: buildbot

Maniphest Tasks: T125

Differential Revision: http://code.redpointsoftware.com.au/D106