Automatically do chmod u+x so Linux works when extracted from ZIP
Summary: Ref T226.  This automatically runs "chmod u+x" when starting Tychaia on Linux.  We need to do this because we have to execute Tychaia.exe again when we want to start the server, but we can't do this unless the executable flag is set.  When extracting from ZIP files, they don't contain executable information (because the executable bit is UNIX-specific and ZIP files originally came from Windows).

Test Plan: Did "chmod u-x Tychaia.exe", ran the game, and saw it get marked with the executable attribute.

Reviewers: Sabre070

Maniphest Tasks: T226

Differential Revision: http://code.redpointsoftware.com.au/D125