Temporary fix for water rendering
Summary: This is a temporary fix for the water rendering so that it doesn't look weird.  We will need to add another flag to Cell to indicate that a cell has water in it, since the blocks with slopes count as "block.Sand", even though we're specially rendering water on them (this means that for game rendering logic, it would detect you as being on land even though you are just on the water's edge).

Test Plan: Ran the game, water didn't look weird.

Reviewers: Sabre070

Reviewed By: Sabre070

Differential Revision: https://code.redpointsoftware.com.au/D150